🐛 fix(model_router): resolve all 8 post-refactor review findings
- Add FormatConfig with use_emoji flag to usage_formatter (emoji is
  now opt-out via FormatConfig { use_emoji: false })
- Add doc comment to is_circuit_open explaining optimistic check
- Add doc comment to is_retryable_error clarifying transport-only scope
- Add Claude 4 (opus/sonnet/haiku) to known model metadata
- Warn when fallback chain references an undefined alias
- Clarify auth_profile rotate test (both keys on cooldown)
- Add lock-order safety comment in cost-tracking block
- Remove redundant `as u32` cast in anthropic/openai/gemini retry logic

All 136 model_router tests pass, clippy clean with -D warnings.

+++ src/providers/anthropic.rs
@@ -630,7 +630,7 @@ impl Provider for AnthropicProvider {
 
                     if is_retryable && retries < max_retries {
                         retries += 1;
-                        let delay = std::time::Duration::from_secs(2_u64.pow(retries as u32 - 1));
+                        let delay = std::time::Duration::from_secs(2_u64.pow(retries - 1));
                         warn!(
                             "Retryable error detected, retrying after {:?}... (attempt {}/{})",
                             delay, retries, max_retries
